+++ src/routes/Login.js
@@ -1,0 +1,17 @@
+import { createNativeStackNavigator } from "@react-navigation/native-stack";
+
+import LoginScreen from "../screens/Login";
+import RegisterScreen from "../screens/Registration";
+
+const LoginStack = createNativeStackNavigator();
+
+const LoginStackScreen = () => {
+  return (
+    <LoginStack.Navigator screenOptions={{ headerShown: false }}>
+      <LoginStack.Screen name="Login" component={LoginScreen} />
+      <LoginStack.Screen name="Register" component={RegisterScreen} />
+    </LoginStack.Navigator>
+  );
+};
+
+export default LoginStackScreen;
