+++ src/routes/Profile.js
@@ -1,0 +1,32 @@
+import { createNativeStackNavigator } from "@react-navigation/native-stack";
+
+import ChangePasswordScreen from "../screens/Profile/ChangePassword";
+import EditProfileScreen from "../screens/Profile/EditProfile";
+import ECContactsScreen from "../screens/Profile/ECContacts";
+import ProfileScreen from "../screens/Profile";
+
+const ProfileStack = createNativeStackNavigator();
+
+const ProfileStackScreen = () => {
+  return (
+    <ProfileStack.Navigator
+      initialRouteName="Profile"
+      screenOptions={{
+        tabBarShowLabel: false,
+        style: {
+          backgroundColor: "#fff",
+        },
+      }}
+    >
+      <ProfileStack.Screen name="Profile" component={ProfileScreen} />
+      <ProfileStack.Screen name="EditProfile" component={EditProfileScreen} />
+      <ProfileStack.Screen name="ECContacts" component={ECContactsScreen} />
+      <ProfileStack.Screen
+        name="ChangePassword"
+        component={ChangePasswordScreen}
+      />
+    </ProfileStack.Navigator>
+  );
+};
+
+export default ProfileStackScreen;
